Running analysis of variance (aov) across multiple grouping variables.

Running analysis of variance (aov) across multiple grouping variables.

grouped_aov(
  data,
  grouping.vars,
  formula,
  effsize = "eta",
  output = "tidy",
  ...
)

Arguments

data
A data frame in which the variables specified in the formula will be found. If missing, the variables are searched for in the standard way.

grouping.vars
Grouping variables.

formula
A formula specifying the model.

effsize
Character describing the effect size to be displayed: "eta" (default) or "omega".

output
A character describing what output is expected. Two possible options: "tidy" (default), which will return the results, or "tukey", which will return results from Tukey's Honest Significant Differences method for post hoc comparisons. The "glance" method to get model summary is currently not supported for this function.

...
Currently ignored.

Examples

# uses dataset included in the `groupedstats` package
set.seed(123)
library(groupedstats)

# effect size
groupedstats::grouped_aov(
  formula = wt ~ mpg,
  data = mtcars,
  grouping.vars = am,
  effsize = "eta"
)

#> # A tibble: 2 x 13
#>      am term  sumsq   df1 meansq F.value   p.value   df2 estimate ci.width
#>   <dbl> <chr> <dbl> <dbl>  <dbl>   <dbl>     <dbl> <dbl>    <dbl>    <dbl>
#> 1     1 mpg    3.77     1   3.77    52.3 0.0000169    11    0.826     0.95
#> 2     0 mpg    6.41     1   6.41    24.4 0.000125     17    0.589     0.95
#>   conf.low conf.high significance
#>      <dbl>     <dbl> <chr>       
#> 1    0.543     0.913 ***         
#> 2    0.243     0.766 ***