Model summary from grouped analysis of any function that has data argument in its function call.

This is the most general form of a grouped_ function where any function can be used to run grouped_ analysis.

grouped_glance(data, grouping.vars, ..f, ...)

Arguments

data
Dataframe from which variables are to be taken.

grouping.vars
List of grouping variables.

..f
A function, or function name as a string.

...
Arguments to function.

These dots support tidy-dots features.

Examples

groupedstats::grouped_glance(
  data = ggplot2::diamonds,
  grouping.vars = c(cut, color),
  formula = price ~ carat * clarity,
  ..f = stats::lm
)

#> # A tibble: 35 x 13
#> # Groups:   cut, color [35]
#>    cut   color r.squared adj.r.squared sigma statistic   p.value    df logLik
#>    <ord> <ord>     <dbl>         <dbl> <dbl>     <dbl>     <dbl> <int>  <dbl>
#>  1 Fair  D         0.915         0.906 1005.      106. 1.03e- 70    16 -1350.
#>  2 Fair  E         0.917         0.912  883.      179. 8.70e-106    14 -1830.
#>  3 Fair  F         0.917         0.912  954.      217. 9.37e-150    16 -2575.
#>  4 Fair  G         0.932         0.929  964.      273. 5.03e-164    16 -2595.
#>  5 Fair  H         0.932         0.929 1033.      332. 1.76e-161    13 -2526.
#>  6 Fair  I         0.958         0.955  794.      307. 1.14e-104    13 -1410.
#>  7 Fair  J         0.955         0.950  907.      204. 1.82e- 66    12  -973.
#>  8 Good  D         0.933         0.931  831.      600. 0.           16 -5382.
#>  9 Good  E         0.927         0.926  905.      781. 0.           16 -7667.
#> 10 Good  F         0.915         0.914  939.      644. 0.           16 -7504.
#> # ... with 25 more rows, and 4 more variables: AIC <dbl>, BIC <dbl>,
#> #   deviance <dbl>, df.residual <int>