Function to run robust simple linear regression (slr) on multiple variables across multiple grouping variables.

Function to run robust simple linear regression (slr) on multiple variables across multiple grouping variables.

grouped_robustslr(data, dep.vars, indep.vars, grouping.vars, nrep = 1000,
  control = robust::lmRob.control(tlo = 1e-04, tua = 1.5e-06, mxr = 50,
  mxf = 50, mxs = 50, tl = 1e-06, estim = "Final", initial.alg = "Auto",
  final.alg = "MM", seed = 1313, level = 0.1, efficiency = 0.9, weight =
  c("Optimal", "Optimal"), trace = TRUE))

Arguments

data
Dataframe from which variables are to be taken.

dep.vars
List criterion or dependent variables for regression (y in y ~ x).

indep.vars
List predictor or independent variables for regression (x in y ~ x).

grouping.vars
List of grouping variables.

nrep
the number of random subsamples to be drawn. If "Exhaustive" resampling is being used, the value of nrep is ignored.

control
a list of control parameters to be used in the numerical algorithms. See lmRob.control for the possible control parameters and their default settings.

Value

A tibble dataframe with tidy results from robust linear regression analyses. The estimates are standardized, i.e. the lm model used is scale(y) ~ scale(x), and not y ~ x.

Examples

# for reproducibility
set.seed(123)

# in case of just one grouping variable
groupedstats::grouped_robustslr(
  data = iris,
  dep.vars = c(Sepal.Length, Petal.Length),
  indep.vars = c(Sepal.Width, Petal.Width),
  grouping.vars = Species
)

#> Warning: Assigning non-quosure objects to quosure lists is deprecated as of rlang 0.3.0.
#> Please coerce to a bare list beforehand with `as.list()`
#> This warning is displayed once per session.

#> # A tibble: 6 x 7
#>   Species   formula            t.value estimate std.error   p.value significance
#>   <fct>     <chr>                <dbl>    <dbl>     <dbl>     <dbl> <chr>       
#> 1 setosa    Sepal.Length ~ Se~    5.39    0.747     0.139   2.12e-6 ***         
#> 2 versicol~ Sepal.Length ~ Se~    3.41    0.526     0.154   1.31e-3 **          
#> 3 virginica Sepal.Length ~ Se~    3.62    0.521     0.144   6.98e-4 ***         
#> 4 setosa    Petal.Length ~ Pe~    2.18    0.301     0.138   3.40e-2 *           
#> 5 versicol~ Petal.Length ~ Pe~    6.75    0.749     0.111   1.77e-8 ***         
#> 6 virginica Petal.Length ~ Pe~    1.23    0.286     0.232   2.23e-1 ns