Function to get descriptive statistics for multiple variables for all grouping variable levels

Function to get descriptive statistics for multiple variables for all grouping variable levels

Arguments

data
Dataframe from which variables need to be taken.

grouping.vars
A list of grouping variables.

measures
List variables for which summary needs to computed. If not spcified, all variables of type specified in the argument measures.type will be used to calculate summaries. Don't explicitly set measures.type = NULL in function call, which will produce an error because the function will try to find a column in a dataframe named "NULL".

measures.type
A character indicating whether summary for numeric ("numeric") or factor/character ("factor") variables is expected (Default: measures.type = "numeric"). This function can't be used for both numeric and variables simultaneously.

topcount.long
If measures.type = factor, you can get the top counts in long format for plotting purposes. (Default: topcount.long = FALSE).

Value

Dataframe with descriptive statistics for numeric variables (n, mean, sd, median, min, max)

Examples

# another possibility
groupedstats::grouped_summary(
data = datasets::iris,
grouping.vars = Species,
measures = Sepal.Length:Petal.Width,
measures.type = "numeric"
)

#> # A tibble: 12 x 13
#>    Species    type    variable     missing complete     n  mean    sd   min
#>    <fct>      <fct>   <fct>          <dbl>    <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 setosa     numeric Petal.Length      0.      50.   50. 1.46  0.170 1.00 
#>  2 setosa     numeric Petal.Width       0.      50.   50. 0.250 0.110 0.100
#>  3 setosa     numeric Sepal.Length      0.      50.   50. 5.01  0.350 4.30 
#>  4 setosa     numeric Sepal.Width       0.      50.   50. 3.43  0.380 2.30 
#>  5 versicolor numeric Petal.Length      0.      50.   50. 4.26  0.470 3.00 
#>  6 versicolor numeric Petal.Width       0.      50.   50. 1.33  0.200 1.00 
#>  7 versicolor numeric Sepal.Length      0.      50.   50. 5.94  0.520 4.90 
#>  8 versicolor numeric Sepal.Width       0.      50.   50. 2.77  0.310 2.00 
#>  9 virginica  numeric Petal.Length      0.      50.   50. 5.55  0.550 4.50 
#> 10 virginica  numeric Petal.Width       0.      50.   50. 2.03  0.270 1.40 
#> 11 virginica  numeric Sepal.Length      0.      50.   50. 6.59  0.640 4.90 
#> 12 virginica  numeric Sepal.Width       0.      50.   50. 2.97  0.320 2.20 
#>      p25 median   p75   max
#>    <dbl>  <dbl> <dbl> <dbl>
#>  1 1.40   1.50  1.58  1.90 
#>  2 0.200  0.200 0.300 0.600
#>  3 4.80   5.00  5.20  5.80 
#>  4 3.20   3.40  3.68  4.40 
#>  5 4.00   4.35  4.60  5.10 
#>  6 1.20   1.30  1.50  1.80 
#>  7 5.60   5.90  6.30  7.00 
#>  8 2.52   2.80  3.00  3.40 
#>  9 5.10   5.55  5.88  6.90 
#> 10 1.80   2.00  2.30  2.50 
#> 11 6.23   6.50  6.90  7.90 
#> 12 2.80   3.00  3.18  3.80 

# if you have just one variable per argument, you can also not use `c()`
groupedstats::grouped_summary(
data = datasets::ToothGrowth,
grouping.vars = supp,
measures = len,
measures.type = "numeric"
)

#> # A tibble: 2 x 13
#>   supp  type    variable missing complete     n  mean    sd   min   p25 median
#>   <fct> <fct>   <fct>      <dbl>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>
#> 1 VC    numeric len           0.      30.   30.  17.0  8.27  4.20  11.2   16.5
#> 2 OJ    numeric len           0.      30.   30.  20.7  6.61  8.20  15.5   22.7
#>     p75   max
#>   <dbl> <dbl>
#> 1  23.1  33.9
#> 2  25.7  30.9