Tidy output from grouped analysis of any function.

This is the most general form of a grouped_ function where any function can be used to run grouped_ analysis.

grouped_tidy(data, grouping.vars, .f, ...)

Arguments

data
Dataframe from which variables are to be taken.

grouping.vars
List of grouping variables.

.f
A function, or function name as a string.

...
Arguments to function.

These dots support tidy-dots features.

Examples

grouped_tidy(
  data = ggplot2::diamonds,
  grouping.vars = c(cut, color),
  formula = price ~ carat * clarity,
  .f = stats::lm
)

#> # A tibble: 547 x 9
#>    cut   color term    estimate std.error statistic   p.value conf.low conf.high
#>    <ord> <ord> <chr>      <dbl>     <dbl>     <dbl>     <dbl>    <dbl>     <dbl>
#>  1 Ideal E     (Inter~  -3085.       64.9   -47.5   0.         -3212.    -2958. 
#>  2 Ideal E     carat    10529.       74.1   142.    0.         10384.    10674. 
#>  3 Ideal E     clarit~  -2088.      267.     -7.81  7.28e- 15  -2612.    -1564. 
#>  4 Ideal E     clarit~    168.      265.      0.633 5.27e-  1   -352.      688. 
#>  5 Ideal E     clarit~   -926.      217.     -4.26  2.06e-  5  -1352.     -500. 
#>  6 Ideal E     clarit~    625.      157.      3.99  6.70e-  5    318.      932. 
#>  7 Ideal E     clarit~   -392.      107.     -3.65  2.63e-  4   -602.     -181. 
#>  8 Ideal E     clarit~     83.9      79.1     1.06  2.89e-  1    -71.1     239. 
#>  9 Ideal E     clarit~    -40.8      67.4    -0.605 5.45e-  1   -173.       91.4
#> 10 Ideal E     carat:~   9746.      287.     34.0   3.38e-222   9185.    10308. 
#> # ... with 537 more rows