Function to run t-test on multiple variables across multiple grouping variables.

Function to run t-test on multiple variables across multiple grouping variables.

grouped_ttest(data, dep.vars, indep.vars, grouping.vars, paired = FALSE,
  var.equal = FALSE)

Arguments

data
Dataframe from which variables are to be taken.

dep.vars
List dependent variables for a t-test (y in y ~ x).

indep.vars
List independent variables for a t-test (x in y ~ x).

grouping.vars
List of grouping variables.

paired
A logical indicating whether you want a paired t-test (Default: paired = FALSE; independent t-test, i.e.).

var.equal
A logical variable indicating whether to treat the two variances as being equal. If TRUE, then the pooled variance is used to estimate the variance otherwise the Welch (or Satterthwaite) approximation to the degrees of freedom is used (Default: var.equal = FALSE; Welch's t-test, i.e.).

Value

A tibble dataframe with tidy results from t-test analyses.

Examples

groupedstats::grouped_ttest(
  data = dplyr::filter(.data = ggplot2::diamonds, color == "E" | color == "J"),
  dep.vars = c(carat, price, depth),
  indep.vars = color,
  grouping.vars = clarity,
  paired = FALSE,
  var.equal = FALSE
)

#> [90m# A tibble: 24 x 10[39m
#>    clarity formula method t.test estimate conf.low conf.high parameter   p.value
#>    [3m[90m<ord>[39m[23m   [3m[90m<chr>[39m[23m   [3m[90m<chr>[39m[23m   [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m
#> [90m 1[39m SI2     carat ~ Welch~ -[31m18[39m[31m.[39m[31m0[39m  -[31m5[39m[31m.[39m[31m0[39m[31m3[39m[90me[39m[31m-1[39m -[31m5[39m[31m.[39m[31m58[39m[90me[39m[31m-1[39m  -[31m4[39m[31m.[39m[31m48[39m[90me[39m[31m-1[39m     634.  9.24[90me[39m[31m- 59[39m
#> [90m 2[39m SI1     carat ~ Welch~ -[31m22[39m[31m.[39m[31m1[39m  -[31m4[39m[31m.[39m[31m62[39m[90me[39m[31m-1[39m -[31m5[39m[31m.[39m[31m0[39m[31m3[39m[90me[39m[31m-1[39m  -[31m4[39m[31m.[39m[31m20[39m[90me[39m[31m-1[39m     948.  2.47[90me[39m[31m- 87[39m
#> [90m 3[39m VS1     carat ~ Welch~ -[31m16[39m[31m.[39m[31m8[39m  -[31m4[39m[31m.[39m[31m44[39m[90me[39m[31m-1[39m -[31m4[39m[31m.[39m[31m96[39m[90me[39m[31m-1[39m  -[31m3[39m[31m.[39m[31m92[39m[90me[39m[31m-1[39m     663.  2.91[90me[39m[31m- 53[39m
#> [90m 4[39m VVS2    carat ~ Welch~ -[31m12[39m[31m.[39m[31m0[39m  -[31m5[39m[31m.[39m[31m53[39m[90me[39m[31m-1[39m -[31m6[39m[31m.[39m[31m44[39m[90me[39m[31m-1[39m  -[31m4[39m[31m.[39m[31m61[39m[90me[39m[31m-1[39m     140.  3.81[90me[39m[31m- 23[39m
#> [90m 5[39m VS2     carat ~ Welch~ -[31m24[39m[31m.[39m[31m6[39m  -[31m5[39m[31m.[39m[31m42[39m[90me[39m[31m-1[39m -[31m5[39m[31m.[39m[31m86[39m[90me[39m[31m-1[39m  -[31m4[39m[31m.[39m[31m99[39m[90me[39m[31m-1[39m     856.  1.51[90me[39m[31m-101[39m
#> [90m 6[39m I1      carat ~ Welch~  -[31m4[39m[31m.[39m[31m48[39m -[31m6[39m[31m.[39m[31m44[39m[90me[39m[31m-1[39m -[31m9[39m[31m.[39m[31m32[39m[90me[39m[31m-1[39m  -[31m3[39m[31m.[39m[31m57[39m[90me[39m[31m-1[39m      60.6 3.34[90me[39m[31m-  5[39m
#> [90m 7[39m VVS1    carat ~ Welch~  -[31m6[39m[31m.[39m[31m53[39m -[31m4[39m[31m.[39m[31m17[39m[90me[39m[31m-1[39m -[31m5[39m[31m.[39m[31m45[39m[90me[39m[31m-1[39m  -[31m2[39m[31m.[39m[31m90[39m[90me[39m[31m-1[39m      76.1 6.68[90me[39m[31m-  9[39m
#> [90m 8[39m IF      carat ~ Welch~  -[31m2[39m[31m.[39m[31m38[39m -[31m1[39m[31m.[39m[31m98[39m[90me[39m[31m-1[39m -[31m3[39m[31m.[39m[31m64[39m[90me[39m[31m-1[39m  -[31m3[39m[31m.[39m[31m18[39m[90me[39m[31m-2[39m      60.1 2.03[90me[39m[31m-  2[39m
#> [90m 9[39m SI2     price ~ Welch~ -[31m10[39m[31m.[39m[31m6[39m  -[31m2[39m[31m.[39m[31m35[39m[90me[39m+3 -[31m2[39m[31m.[39m[31m78[39m[90me[39m+3  -[31m1[39m[31m.[39m[31m91[39m[90me[39m+3     676.  2.02[90me[39m[31m- 24[39m
#> [90m10[39m SI1     price ~ Welch~ -[31m12[39m[31m.[39m[31m6[39m  -[31m2[39m[31m.[39m[31m0[39m[31m2[39m[90me[39m+3 -[31m2[39m[31m.[39m[31m34[39m[90me[39m+3  -[31m1[39m[31m.[39m[31m71[39m[90me[39m+3    [4m1[24m031.  5.75[90me[39m[31m- 34[39m
#> [90m# ... with 14 more rows, and 1 more variable: significance [3m[90m<chr>[90m[23m[39m