Function to run two-sample Wilcoxon tests on multiple variables across multiple grouping variables.

Function to run two-sample Wilcoxon tests on multiple variables across multiple grouping variables.

grouped_wilcox(data, dep.vars, indep.vars, grouping.vars, paired = FALSE,
  correct = TRUE)

Arguments

data
Dataframe from which variables are to be taken.

dep.vars
List dependent variables for a two-sample Wilcoxon tests (y in y ~ x).

indep.vars
List independent variables for a two-sample Wilcoxon tests (x in y ~ x).

grouping.vars
List of grouping variables.

paired
A logical indicating whether you want a paired two-sample Wilcoxon tests (Default: paired = FALSE).

correct
A logical indicating whether to apply continuity correction in the normal approximation for the p-value (Default: correct = TRUE).

Value

A tibble dataframe with tidy results from two-sample Wilcoxon tests analyses.

Examples

# only with one grouping variable
groupedstats::grouped_wilcox(
  data = dplyr::filter(.data = ggplot2::diamonds, color == "E" | color == "J"),
  dep.vars = depth:table,
  indep.vars = color,
  grouping.vars = clarity,
  paired = FALSE
)

#> [90m# A tibble: 16 x 9[39m
#>    clarity formula method statistic  estimate  conf.low conf.high  p.value
#>    [3m[90m<ord>[39m[23m   [3m[90m<chr>[39m[23m   [3m[90m<chr>[39m[23m      [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m
#> [90m 1[39m SI2     depth ~ Wilco~   [4m3[24m[4m8[24m[4m0[24m600   -[31m2[39m[31m.[39m[31m00[39m[90me[39m[31m-1[39m  -[31m3[39m[31m.[39m[31m00[39m[90me[39m[31m-1[39m  -[31m3[39m[31m.[39m[31m64[39m[90me[39m[31m-5[39m 1.54[90me[39m[31m- 2[39m
#> [90m 2[39m SI1     depth ~ Wilco~   [4m9[24m[4m1[24m[4m8[24m892    3.93[90me[39m[31m-5[39m -[31m10[39m[31m.[39m[31m00[39m[90me[39m[31m-2[39m   1.00[90me[39m[31m-1[39m 6.77[90me[39m[31m- 1[39m
#> [90m 3[39m VS1     depth ~ Wilco~   [4m2[24m[4m7[24m[4m6[24m743   -[31m4[39m[31m.[39m[31m00[39m[90me[39m[31m-1[39m  -[31m5[39m[31m.[39m[31m00[39m[90me[39m[31m-1[39m  -[31m3[39m[31m.[39m[31m00[39m[90me[39m[31m-1[39m 7.03[90me[39m[31m-12[39m
#> [90m 4[39m VVS2    depth ~ Wilco~    [4m5[24m[4m2[24m618   -[31m4[39m[31m.[39m[31m00[39m[90me[39m[31m-1[39m  -[31m6[39m[31m.[39m[31m00[39m[90me[39m[31m-1[39m  -[31m2[39m[31m.[39m[31m00[39m[90me[39m[31m-1[39m 4.18[90me[39m[31m- 4[39m
#> [90m 5[39m VS2     depth ~ Wilco~   [4m8[24m[4m1[24m[4m6[24m784.  -[31m2[39m[31m.[39m[31m00[39m[90me[39m[31m-1[39m  -[31m3[39m[31m.[39m[31m00[39m[90me[39m[31m-1[39m -[31m10[39m[31m.[39m[31m00[39m[90me[39m[31m-2[39m 8.86[90me[39m[31m- 5[39m
#> [90m 6[39m I1      depth ~ Wilco~     [4m1[24m570.  -[31m2[39m[31m.[39m[31m00[39m[90me[39m+0  -[31m3[39m[31m.[39m[31m00[39m[90me[39m+0 -[31m10[39m[31m.[39m[31m00[39m[90me[39m[31m-1[39m 1.21[90me[39m[31m- 4[39m
#> [90m 7[39m VVS1    depth ~ Wilco~    [4m1[24m[4m9[24m455   -[31m3[39m[31m.[39m[31m00[39m[90me[39m[31m-1[39m  -[31m5[39m[31m.[39m[31m00[39m[90me[39m[31m-1[39m -[31m10[39m[31m.[39m[31m00[39m[90me[39m[31m-2[39m 5.07[90me[39m[31m- 3[39m
#> [90m 8[39m IF      depth ~ Wilco~     [4m3[24m286   -[31m3[39m[31m.[39m[31m00[39m[90me[39m[31m-1[39m  -[31m7[39m[31m.[39m[31m00[39m[90me[39m[31m-1[39m  -[31m2[39m[31m.[39m[31m51[39m[90me[39m[31m-5[39m 4.79[90me[39m[31m- 2[39m
#> [90m 9[39m SI2     table ~ Wilco~   [4m3[24m[4m8[24m[4m8[24m138.  -[31m2[39m[31m.[39m[31m0[39m[31m8[39m[90me[39m[31m-5[39m  -[31m3[39m[31m.[39m[31m91[39m[90me[39m[31m-5[39m   4.98[90me[39m[31m-6[39m 6.85[90me[39m[31m- 2[39m
#> [90m10[39m SI1     table ~ Wilco~   [4m8[24m[4m8[24m[4m0[24m016.   0.  [90m [39m    -[31m1[39m[31m.[39m[31m30[39m[90me[39m[31m-5[39m   3.62[90me[39m[31m-5[39m 1.71[90me[39m[31m- 1[39m
#> [90m11[39m VS1     table ~ Wilco~   [4m3[24m[4m0[24m[4m6[24m400   -[31m4[39m[31m.[39m[31m82[39m[90me[39m[31m-6[39m -[31m10[39m[31m.[39m[31m00[39m[90me[39m[31m-1[39m  -[31m2[39m[31m.[39m[31m87[39m[90me[39m[31m-5[39m 6.16[90me[39m[31m- 5[39m
#> [90m12[39m VVS2    table ~ Wilco~    [4m5[24m[4m4[24m524  -[31m10[39m[31m.[39m[31m00[39m[90me[39m[31m-1[39m  -[31m1[39m[31m.[39m[31m00[39m[90me[39m+0  -[31m5[39m[31m.[39m[31m87[39m[90me[39m[31m-5[39m 2.60[90me[39m[31m- 3[39m
#> [90m13[39m VS2     table ~ Wilco~   [4m7[24m[4m6[24m[4m3[24m942. -[31m10[39m[31m.[39m[31m00[39m[90me[39m[31m-1[39m -[31m10[39m[31m.[39m[31m00[39m[90me[39m[31m-1[39m  -[31m4[39m[31m.[39m[31m61[39m[90me[39m[31m-5[39m 1.65[90me[39m[31m-10[39m
#> [90m14[39m I1      table ~ Wilco~     [4m2[24m900   10.00[90me[39m[31m-1[39m  -[31m2[39m[31m.[39m[31m80[39m[90me[39m[31m-5[39m   2.00[90me[39m+0 1.68[90me[39m[31m- 1[39m
#> [90m15[39m VVS1    table ~ Wilco~    [4m2[24m[4m2[24m659   -[31m7[39m[31m.[39m[31m72[39m[90me[39m[31m-6[39m -[31m10[39m[31m.[39m[31m00[39m[90me[39m[31m-1[39m   4.05[90me[39m[31m-6[39m 3.42[90me[39m[31m- 1[39m
#> [90m16[39m IF      table ~ Wilco~     [4m3[24m987   -[31m3[39m[31m.[39m[31m35[39m[90me[39m[31m-5[39m -[31m10[39m[31m.[39m[31m00[39m[90me[39m[31m-1[39m  10.00[90me[39m[31m-1[39m 9.11[90me[39m[31m- 1[39m
#> [90m# ... with 1 more variable: significance [3m[90m<chr>[90m[23m[39m