Confidence intervals for partial eta-squared and omega-squared for linear models.

This function will convert a linear model object to a dataframe containing statistical details for all effects along with partial eta-squared effect size and its confidence interval.

lm_effsize_ci(object, effsize = "eta", partial = TRUE,
  conf.level = 0.95, nboot = 500)

Arguments

object
The linear model object (can be of class lm, aov, anova, or aovlist).

effsize
Character describing the effect size to be displayed: "eta" (default) or "omega".

partial
Logical that decides if partial eta-squared or omega-squared are returned (Default: TRUE). If FALSE, eta-squared or omega-squared will be returned. Valid only for objects of class lm, aov, anova, or aovlist.

conf.level
Numeric specifying Level of confidence for the confidence interval (Default: 0.95).

nboot
Number of bootstrap samples for confidence intervals for partial eta-squared and omega-squared (Default: 500).

Value

A dataframe with results from stats::lm() with partial eta-squared, omega-squared, and bootstrapped confidence interval for the same.

Examples

# model
set.seed(123)
mod <-
  stats::aov(
    formula = mpg ~ wt + qsec + Error(disp / am),
    data = mtcars
  )

# dataframe with effect size and confidence intervals
groupedstats::lm_effsize_ci(mod)

#> # A tibble: 2 x 9
#>   term  F.value   df1   df2 p.value partial.etasq conf.low conf.high stratum
#>   <chr>   <dbl> <dbl> <dbl>   <dbl>         <dbl>    <dbl>     <dbl> <chr>  
#> 1 wt      12.0      1    27 0.00176         0.308   0.0533     0.520 Within 
#> 2 qsec     4.72     1    27 0.0387          0.149   0          0.378 Within