Standardize a dataframe with effect sizes for aov, lm, aovlist, etc. objects.

The difference between lm_effsize_ci and lm_effsize_standardizer is that the former has more opinionated column naming, while the latter doesn't. The latter can thus be more helpful in writing a wrapper around this function.

lm_effsize_standardizer(
  object,
  effsize = "eta",
  partial = TRUE,
  conf.level = 0.95,
  ...
)

Arguments

object
The linear model object (can be of class lm, aov, anova, or aovlist).

effsize
Character describing the effect size to be displayed: "eta" (default) or "omega".

partial
Logical that decides if partial eta-squared or omega-squared are returned (Default: TRUE). If FALSE, eta-squared or omega-squared will be returned. Valid only for objects of class lm, aov, anova, or aovlist.

conf.level
Numeric specifying Level of confidence for the confidence interval (Default: 0.95).

...
Ignored.

Examples

set.seed(123)
groupedstats::lm_effsize_standardizer(
  object = stats::lm(formula = brainwt ~ vore, data = ggplot2::msleep),
  effsize = "eta",
  partial = FALSE,
  conf.level = 0.99
)

#> # A tibble: 1 x 9
#>   term  F.value   df1   df2 p.value estimate conf.level conf.low conf.high
#>   <chr>   <dbl> <int> <int>   <dbl>    <dbl>      <dbl>    <dbl>     <dbl>
#> 1 vore     1.06     3    47   0.375   0.0633       0.99        0     0.256