Running analysis of variance (aov) across multiple grouping variables.

Running analysis of variance (aov) across multiple grouping variables.

grouped_aov(
  data,
  grouping.vars,
  formula,
  effsize = "eta",
  output = "tidy",
  nboot = 1000
)

Arguments

data
A data frame in which the variables specified in the formula will be found. If missing, the variables are searched for in the standard way.

grouping.vars
Grouping variables.

formula
A formula specifying the model.

effsize
Character describing the effect size to be displayed: "eta" (default) or "omega".

output
A character describing what output is expected. Two possible options: "tidy" (default), which will return the results, or "tukey", which will return results from Tukey's Honest Significant Differences method for post hoc comparisons. The "glance" method to get model summary is currently not supported for this function.

nboot
Number of bootstrap samples for confidence intervals for partial eta-squared and omega-squared (Default: 500).

Value

A tibble dataframe with tidy results from anova. No model summaries available.

Examples

# uses dataset included in the groupedstats package
set.seed(123)
library(groupedstats)

groupedstats::grouped_aov(
  formula = rating ~ belief * outcome * question,
  data = intent_morality,
  grouping.vars = item,
  effsize = "eta"
)

#> # A tibble: 28 x 10
#>    item  term  F.value   df1   df2   p.value partial.etasq conf.low conf.high
#>    <fct> <chr>   <dbl> <dbl> <dbl>     <dbl>         <dbl>    <dbl>     <dbl>
#>  1 Grace beli~ 2233.       1   996 1.21e-256       0.692   0.664       0.716 
#>  2 Grace outc~   47.7      1   996 8.69e- 12       0.0457  0.0237      0.0734
#>  3 Grace ques~   17.3      1   996 3.47e-  5       0.0171  0.00478     0.0362
#>  4 Grace beli~   28.5      1   996 1.16e-  7       0.0278  0.0112      0.0508
#>  5 Grace beli~    6.14     1   996 1.34e-  2       0.00613 0.000238    0.0194
#>  6 Grace outc~    8.62     1   996 3.40e-  3       0.00858 0.000942    0.0235
#>  7 Grace beli~    7.91     1   996 5.00e-  3       0.00788 0.000717    0.0223
#>  8 Anna  beli~  331.       1   996 4.85e- 64       0.249   0.205       0.292 
#>  9 Anna  outc~   86.2      1   996 9.68e- 20       0.0797  0.0505      0.113 
#> 10 Anna  ques~   47.7      1   996 8.76e- 12       0.0457  0.0237      0.0734
#> # ... with 18 more rows, and 1 more variable: significance <chr>