Running analysis of variance (aov) across multiple grouping variables.

Running analysis of variance (aov) across multiple grouping variables.

grouped_aov(
  data,
  grouping.vars,
  formula,
  effsize = "eta",
  output = "tidy",
  ...
)

Arguments

data
A data frame in which the variables specified in the formula will be found. If missing, the variables are searched for in the standard way.

grouping.vars
Grouping variables.

formula
A formula specifying the model.

effsize
Character describing the effect size to be displayed: "eta" (default) or "omega".

output
A character describing what output is expected. Two possible options: "tidy" (default), which will return the results, or "tukey", which will return results from Tukey's Honest Significant Differences method for post hoc comparisons. The "glance" method to get model summary is currently not supported for this function.

...
Currently ignored.

Examples

# uses dataset included in the `groupedstats` package
set.seed(123)
library(groupedstats)

# effect size
groupedstats::grouped_aov(
  formula = rating ~ belief * outcome * question,
  data = intent_morality,
  grouping.vars = item,
  effsize = "eta"
)

#> # A tibble: 28 x 11
#>    item  term                    F.value   df1   df2   p.value eta.sq.partial
#>    <fct> <chr>                     <dbl> <dbl> <dbl>     <dbl>          <dbl>
#>  1 Grace belief                  2233.       1   996 1.21e-256        0.692  
#>  2 Grace outcome                   47.7      1   996 8.69e- 12        0.0457 
#>  3 Grace question                  17.3      1   996 3.47e-  5        0.0171 
#>  4 Grace belief:outcome            28.5      1   996 1.16e-  7        0.0278 
#>  5 Grace belief:question            6.14     1   996 1.34e-  2        0.00613
#>  6 Grace outcome:question           8.62     1   996 3.40e-  3        0.00858
#>  7 Grace belief:outcome:question    7.91     1   996 5.00e-  3        0.00788
#>  8 Anna  belief                   331.       1   996 4.85e- 64        0.249  
#>  9 Anna  outcome                   86.2      1   996 9.68e- 20        0.0797 
#> 10 Anna  question                  47.7      1   996 8.76e- 12        0.0457 
#>    conf.level conf.low conf.high significance
#>         <dbl>    <dbl>     <dbl> <chr>       
#>  1       0.95 0.597       0.721  ***         
#>  2       0.95 0.0238      0.0735 ***         
#>  3       0.95 0.00479     0.0363 ***         
#>  4       0.95 0.0112      0.0509 ***         
#>  5       0.95 0.000239    0.0194 *           
#>  6       0.95 0.000944    0.0235 **          
#>  7       0.95 0.000719    0.0224 **          
#>  8       0.95 0.206       0.293  ***         
#>  9       0.95 0.0506      0.113  ***         
#> 10       0.95 0.0238      0.0735 ***         
#> # ... with 18 more rows

# pairwise comparisons
groupedstats::grouped_aov(
  formula = rating ~ belief * outcome * question,
  data = intent_morality,
  grouping.vars = item,
  output = "tukey"
)

#> Note: The p-value is adjusted for the number of tests conducted
#>             at each level of the grouping variable.

#> # A tibble: 196 x 8
#>    item  term            comparison                             estimate
#>    <fct> <chr>           <chr>                                     <dbl>
#>  1 Grace belief          neutral-negative                         -4.37 
#>  2 Grace outcome         neutral-negative                         -0.639
#>  3 Grace question        wrongness-punishment                      0.384
#>  4 Grace belief:outcome  neutral:negative-negative:negative       -4.86 
#>  5 Grace belief:outcome  negative:neutral-negative:negative       -1.13 
#>  6 Grace belief:outcome  neutral:neutral-negative:negative        -5.01 
#>  7 Grace belief:outcome  negative:neutral-neutral:negative         3.73 
#>  8 Grace belief:outcome  neutral:neutral-neutral:negative         -0.145
#>  9 Grace belief:outcome  neutral:neutral-negative:neutral         -3.88 
#> 10 Grace belief:question neutral:punishment-negative:punishment   -4.14 
#>    conf.low conf.high adj.p.value significance
#>       <dbl>     <dbl>       <dbl> <chr>       
#>  1   -4.55     -4.19     0.       ***         
#>  2   -0.820    -0.457    8.48e-12 ***         
#>  3    0.203     0.566    3.47e- 5 ***         
#>  4   -5.20     -4.53     0.       ***         
#>  5   -1.47     -0.796    0.       ***         
#>  6   -5.35     -4.67     0.       ***         
#>  7    3.40      4.07     0.       ***         
#>  8   -0.482     0.191    6.83e- 1 ns          
#>  9   -4.21     -3.54     0.       ***         
#> 10   -4.48     -3.80     0.       ***         
#> # ... with 186 more rows