Function to run generalized linear model (glm) across multiple grouping variables.

Function to run generalized linear model (glm) across multiple grouping variables.

grouped_glm(data, grouping.vars, formula, family = stats::binomial(link =
  "logit"), quick = FALSE, exponentiate = FALSE, output = "tidy")

Arguments

data
Dataframe from which variables are to be taken.

grouping.vars
List of grouping variables.

formula
an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted. The details of model specification are given under ‘Details’.

family
a description of the error distribution and link function to be used in the model. For glm this can be a character string naming a family function, a family function or the result of a call to a family function. For glm.fit only the third option is supported. (See family for details of family functions.)

quick
Logical indicating if the only the term and estimate columns should be returned. Often useful to avoid time consuming covariance and standard error calculations. Defaults to FALSE.

exponentiate
Logical indicating whether or not to exponentiate the the coefficient estimates. This is typical for logistic and multinomial regressions, but a bad idea if there is no log or logit link. Defaults to FALSE.

output
A character describing what output is expected. Two possible options: "tidy" (default), which will return the results, or "glance", which will return model summaries.

Value

A tibble dataframe with tidy results from linear model.

See also

Examples

# to get tidy output
groupedstats::grouped_glm(
  data = groupedstats::Titanic_full,
  formula = Survived ~ Sex,
  grouping.vars = Class,
  family = stats::binomial(link = "logit")
)

#> [90m# A tibble: 8 x 9[39m
#>   Class term  estimate std.error statistic conf.low conf.high  p.value
#>   [3m[90m<fct>[39m[23m [3m[90m<chr>[39m[23m    [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m
#> [90m1[39m 3rd   (Int~   -[31m0[39m[31m.[39m[31m164[39m     0.143     -[31m1[39m[31m.[39m[31m14[39m   -[31m0[39m[31m.[39m[31m446[39m     0.117 2.54[90me[39m[31m- 1[39m
#> [90m2[39m 3rd   SexM~   -[31m1[39m[31m.[39m[31m40[39m      0.185     -[31m7[39m[31m.[39m[31m58[39m   -[31m1[39m[31m.[39m[31m77[39m     -[31m1[39m[31m.[39m[31m0[39m[31m4[39m  3.36[90me[39m[31m-14[39m
#> [90m3[39m 1st   (Int~    3.56      0.507      7.03    2.70      4.74  2.13[90me[39m[31m-12[39m
#> [90m4[39m 1st   SexM~   -[31m4[39m[31m.[39m[31m21[39m      0.531     -[31m7[39m[31m.[39m[31m92[39m   -[31m5[39m[31m.[39m[31m42[39m     -[31m3[39m[31m.[39m[31m29[39m  2.29[90me[39m[31m-15[39m
#> [90m5[39m 2nd   (Int~    1.97      0.296      6.65    1.43      2.60  3.03[90me[39m[31m-11[39m
#> [90m6[39m 2nd   SexM~   -[31m3[39m[31m.[39m[31m79[39m      0.366    -[31m10[39m[31m.[39m[31m3[39m    -[31m4[39m[31m.[39m[31m54[39m     -[31m3[39m[31m.[39m[31m10[39m  4.88[90me[39m[31m-25[39m
#> [90m7[39m Crew  (Int~    1.90      0.619      3.06    0.827     3.34  2.18[90me[39m[31m- 3[39m
#> [90m8[39m Crew  SexM~   -[31m3[39m[31m.[39m[31m15[39m      0.625     -[31m5[39m[31m.[39m[31m0[39m[31m4[39m   -[31m4[39m[31m.[39m[31m60[39m     -[31m2[39m[31m.[39m[31m0[39m[31m6[39m  4.68[90me[39m[31m- 7[39m
#> [90m# ... with 1 more variable: significance [3m[90m<chr>[90m[23m[39m

# to get glance output
groupedstats::grouped_glm(
  data = groupedstats::Titanic_full,
  formula = Survived ~ Sex,
  grouping.vars = Class,
  family = stats::binomial(link = "logit"),
  output = "glance"
)

#> [90m# A tibble: 4 x 8[39m
#>   Class null.deviance df.null logLik   AIC   BIC deviance df.residual
#>   [3m[90m<fct>[39m[23m         [3m[90m<dbl>[39m[23m   [3m[90m<int>[39m[23m  [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m       [3m[90m<int>[39m[23m
#> [90m1[39m 3rd            797.     705  -[31m370[39m[31m.[39m  744.  753.     740.         704
#> [90m2[39m 1st            430.     324  -[31m134[39m[31m.[39m  272.  280.     268.         323
#> [90m3[39m 2nd            387.     284  -[31m112[39m[31m.[39m  228.  235.     224.         283
#> [90m4[39m Crew           974.     884  -[31m466[39m[31m.[39m  936.  946.     932.         883