Function to run generalized linear model (glm) across multiple grouping variables.

Function to run generalized linear model (glm) across multiple grouping variables.

grouped_glm(
  data,
  grouping.vars,
  ...,
  output = "tidy",
  tidy.args = list(conf.int = TRUE, conf.level = 0.95),
  augment.args = list()
)

Arguments

data
Dataframe (or tibble) from which variables are to be taken.

grouping.vars
Grouping variables.

...
Additional arguments to broom::tidy, broom::glance, or broom::augment S3 method.

output
A character describing what output is expected. Two possible options: "tidy" (default), which will return the results, or "glance", which will return model summaries.

tidy.args
A list of arguments to be used in the relevant S3 method.

augment.args
A list of arguments to be used in the relevant S3 method.

Value

A tibble dataframe with tidy results from linear model.

See also

Examples

# to get tidy output
groupedstats::grouped_glm(
  data = groupedstats::Titanic_full,
  formula = Survived ~ Sex,
  grouping.vars = Class,
  family = stats::binomial(link = "logit")
)

#> # A tibble: 8 x 9
#>   Class term        estimate std.error statistic  p.value conf.low conf.high
#>   <fct> <chr>          <dbl>     <dbl>     <dbl>    <dbl>    <dbl>     <dbl>
#> 1 1st   (Intercept)    3.56      0.507      7.03 2.13e-12    2.70      4.74 
#> 2 1st   SexMale       -4.21      0.531     -7.92 2.29e-15   -5.42     -3.29 
#> 3 2nd   (Intercept)    1.97      0.296      6.65 3.03e-11    1.43      2.60 
#> 4 2nd   SexMale       -3.79      0.366    -10.3  4.88e-25   -4.54     -3.10 
#> 5 3rd   (Intercept)   -0.164     0.143     -1.14 2.54e- 1   -0.446     0.117
#> 6 3rd   SexMale       -1.40      0.185     -7.58 3.36e-14   -1.77     -1.04 
#> 7 Crew  (Intercept)    1.90      0.619      3.06 2.18e- 3    0.827     3.34 
#> 8 Crew  SexMale       -3.15      0.625     -5.04 4.68e- 7   -4.60     -2.06 
#>   significance
#>   <chr>       
#> 1 ***         
#> 2 ***         
#> 3 ***         
#> 4 ***         
#> 5 ns          
#> 6 ***         
#> 7 **          
#> 8 ***         

# to get glance output
groupedstats::grouped_glm(
  data = groupedstats::Titanic_full,
  formula = Survived ~ Sex,
  grouping.vars = Class,
  family = stats::binomial(link = "logit"),
  output = "glance"
)

#> # A tibble: 4 x 9
#>   Class null.deviance df.null logLik   AIC   BIC deviance df.residual  nobs
#>   <fct>         <dbl>   <int>  <dbl> <dbl> <dbl>    <dbl>       <int> <int>
#> 1 1st            430.     324  -134.  272.  280.     268.         323   325
#> 2 2nd            387.     284  -112.  228.  235.     224.         283   285
#> 3 3rd            797.     705  -370.  744.  753.     740.         704   706
#> 4 Crew           974.     884  -466.  936.  946.     932.         883   885