Function to run generalized linear model (glm) across multiple grouping variables.

Function to run generalized linear model (glm) across multiple grouping variables.

grouped_glm(data, grouping.vars, formula, family = stats::binomial(link =
  "logit"), quick = FALSE, exponentiate = FALSE, output = "tidy")

Arguments

data
Dataframe from which variables are to be taken.

grouping.vars
List of grouping variables.

formula
an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted. The details of model specification are given under ‘Details’.

family
a description of the error distribution and link function to be used in the model. For glm this can be a character string naming a family function, a family function or the result of a call to a family function. For glm.fit only the third option is supported. (See family for details of family functions.)

quick
Logical indicating if the only the term and estimate columns should be returned. Often useful to avoid time consuming covariance and standard error calculations. Defaults to FALSE.

exponentiate
Logical indicating whether or not to exponentiate the the coefficient estimates. This is typical for logistic and multinomial regressions, but a bad idea if there is no log or logit link. Defaults to FALSE.

output
A character describing what output is expected. Two possible options: "tidy" (default), which will return the results, or "glance", which will return model summaries.

Value

A tibble dataframe with tidy results from linear model.

See also

Examples

groupedstats::grouped_glm(
  data = ggstatsplot::Titanic_full,
  formula = Survived ~ Sex,
  grouping.vars = Class,
  family = stats::binomial(link = "logit")
)

#> # A tibble: 8 x 9
#>   Class term  estimate std.error statistic conf.low conf.high  p.value
#>   <fct> <chr>    <dbl>     <dbl>     <dbl>    <dbl>     <dbl>    <dbl>
#> 1 3rd   (Int~   -0.164     0.143     -1.14   -0.446     0.117 2.54e- 1
#> 2 3rd   SexM~   -1.40      0.185     -7.58   -1.77     -1.04  3.36e-14
#> 3 1st   (Int~    3.56      0.507      7.03    2.70      4.74  2.13e-12
#> 4 1st   SexM~   -4.21      0.531     -7.92   -5.42     -3.29  2.29e-15
#> 5 2nd   (Int~    1.97      0.296      6.65    1.43      2.60  3.03e-11
#> 6 2nd   SexM~   -3.79      0.366    -10.3    -4.54     -3.10  4.88e-25
#> 7 Crew  (Int~    1.90      0.619      3.06    0.827     3.34  2.18e- 3
#> 8 Crew  SexM~   -3.15      0.625     -5.04   -4.60     -2.06  4.68e- 7
#> # ... with 1 more variable: significance <chr>