Function to run generalized linear model (glm) across multiple grouping variables.

Function to run generalized linear model (glm) across multiple grouping variables.

grouped_glm(data, grouping.vars, ..., output = "tidy",
  tidy.args = list(conf.int = TRUE, conf.level = 0.95),
  augment.args = list())

Arguments

data
Dataframe (or tibble) from which variables are to be taken.

grouping.vars
Grouping variables.

...
Arguments to function.

These dots support tidy-dots features.

output
A character describing what output is expected. Two possible options: "tidy" (default), which will return the results, or "glance", which will return model summaries.

tidy.args
A list of arguments to be used in the relevant S3 method.

augment.args
A list of arguments to be used in the relevant S3 method.

Value

A tibble dataframe with tidy results from linear model.

See also

Examples

# to get tidy output
groupedstats::grouped_glm(
  data = groupedstats::Titanic_full,
  formula = Survived ~ Sex,
  grouping.vars = Class,
  family = stats::binomial(link = "logit")
)

#> # A tibble: 8 x 9
#>   Class term  estimate std.error statistic conf.low conf.high  p.value
#>   <fct> <chr>    <dbl>     <dbl>     <dbl>    <dbl>     <dbl>    <dbl>
#> 1 1st   (Int~    3.56      0.507      7.03    2.70      4.74  2.13e-12
#> 2 1st   SexM~   -4.21      0.531     -7.92   -5.42     -3.29  2.29e-15
#> 3 2nd   (Int~    1.97      0.296      6.65    1.43      2.60  3.03e-11
#> 4 2nd   SexM~   -3.79      0.366    -10.3    -4.54     -3.10  4.88e-25
#> 5 3rd   (Int~   -0.164     0.143     -1.14   -0.446     0.117 2.54e- 1
#> 6 3rd   SexM~   -1.40      0.185     -7.58   -1.77     -1.04  3.36e-14
#> 7 Crew  (Int~    1.90      0.619      3.06    0.827     3.34  2.18e- 3
#> 8 Crew  SexM~   -3.15      0.625     -5.04   -4.60     -2.06  4.68e- 7
#> # ... with 1 more variable: significance <chr>

# to get glance output
groupedstats::grouped_glm(
  data = groupedstats::Titanic_full,
  formula = Survived ~ Sex,
  grouping.vars = Class,
  family = stats::binomial(link = "logit"),
  output = "glance"
)

#> # A tibble: 4 x 9
#>   Class null.deviance df.null logLik   AIC   BIC deviance df.residual  nobs
#>   <fct>         <dbl>   <int>  <dbl> <dbl> <dbl>    <dbl>       <int> <int>
#> 1 1st            430.     324  -134.  272.  280.     268.         323   325
#> 2 2nd            387.     284  -112.  228.  235.     224.         283   285
#> 3 3rd            797.     705  -370.  744.  753.     740.         704   706
#> 4 Crew           974.     884  -466.  936.  946.     932.         883   885