Function to run generalized linear mixed-effects model (glmer) across multiple grouping variables.

Function to run generalized linear mixed-effects model (glmer) across multiple grouping variables.

grouped_glmer(data, grouping.vars, formula, family = stats::binomial(link
  = "probit"), control = lme4::glmerControl(optimizer = "bobyqa",
  boundary.tol = 1e-07, calc.derivs = FALSE, use.last.params = FALSE,
  optCtrl = list(maxfun = 2e+09)), output = "tidy")

Arguments

data
Dataframe from which variables are to be taken.

grouping.vars
List of grouping variables.

formula
a two-sided linear formula object describing both the fixed-effects and random-effects part of the model, with the response on the left of a ~ operator and the terms, separated by + operators, on the right. Random-effects terms are distinguished by vertical bars ("|") separating expressions for design matrices from grouping factors.

family
a GLM family, see glm and family.

control
a list (of correct class, resulting from lmerControl() or glmerControl() respectively) containing control parameters, including the nonlinear optimizer to be used and parameters to be passed through to the nonlinear optimizer, see the *lmerControl documentation for details.

output
A character describing what output is expected. Two possible options: "tidy" (default), which will return the results, or "glance", which will return model summaries.

Value

A tibble dataframe with tidy results from linear model or model summaries.

See also

Examples

# commented out because the examples are time-consuming and the R CMD CHECK
# makes a NOTE (> 5s)

# categorical outcome; binomial family
groupedstats::grouped_glmer(
  formula = Survived ~ Age + (Age |
    Class),
  family = stats::binomial(link = "probit"),
  data = dplyr::sample_frac(groupedstats::Titanic_full, size = 0.3),
  grouping.vars = Sex
)

#> # A tibble: 4 x 10
#>   Sex   effect term  estimate std.error statistic conf.low conf.high  p.value
#>   <fct> <chr>  <chr>    <dbl>     <dbl>     <dbl>    <dbl>     <dbl>    <dbl>
#> 1 Male  fixed  (Int~   -0.914     0.141     -6.50 -1.19       -0.639 8.25e-11
#> 2 Male  fixed  AgeC~    2.25      1.07       2.10  0.151       4.34  3.56e- 2
#> 3 Fema~ fixed  (Int~    0.841     0.426      1.97  0.00600     1.68  4.84e- 2
#> 4 Fema~ fixed  AgeC~    0.687     0.548      1.25 -0.388       1.76  2.10e- 1
#> # ... with 1 more variable: significance <chr>

# continuous outcome; gaussian family
library(gapminder)

groupedstats::grouped_glmer(
  data = dplyr::sample_frac(gapminder, size = 0.3),
  formula = scale(lifeExp) ~ scale(gdpPercap) + (gdpPercap | continent),
  family = stats::gaussian(),
  control = lme4::lmerControl(
    optimizer = "bobyqa",
    restart_edge = TRUE,
    boundary.tol = 1e-7,
    calc.derivs = FALSE,
    optCtrl = list(maxfun = 2e9)
  ),
  grouping.vars = year,
  output = "tidy"
)

#> Warning: please use glmerControl() instead of lmerControl()

#> Warning: calling glmer() with family=gaussian (identity link) as a shortcut to lmer() is deprecated; please call lmer() directly

#> Warning: please use glmerControl() instead of lmerControl()

#> Warning: calling glmer() with family=gaussian (identity link) as a shortcut to lmer() is deprecated; please call lmer() directly

#> Warning: please use glmerControl() instead of lmerControl()

#> Warning: calling glmer() with family=gaussian (identity link) as a shortcut to lmer() is deprecated; please call lmer() directly

#> Warning: please use glmerControl() instead of lmerControl()

#> Warning: calling glmer() with family=gaussian (identity link) as a shortcut to lmer() is deprecated; please call lmer() directly

#> Warning: please use glmerControl() instead of lmerControl()

#> Warning: calling glmer() with family=gaussian (identity link) as a shortcut to lmer() is deprecated; please call lmer() directly

#> Warning: please use glmerControl() instead of lmerControl()

#> Warning: calling glmer() with family=gaussian (identity link) as a shortcut to lmer() is deprecated; please call lmer() directly

#> Warning: please use glmerControl() instead of lmerControl()

#> Warning: calling glmer() with family=gaussian (identity link) as a shortcut to lmer() is deprecated; please call lmer() directly

#> Warning: please use glmerControl() instead of lmerControl()

#> Warning: calling glmer() with family=gaussian (identity link) as a shortcut to lmer() is deprecated; please call lmer() directly

#> Warning: please use glmerControl() instead of lmerControl()

#> Warning: calling glmer() with family=gaussian (identity link) as a shortcut to lmer() is deprecated; please call lmer() directly

#> Warning: please use glmerControl() instead of lmerControl()

#> Warning: calling glmer() with family=gaussian (identity link) as a shortcut to lmer() is deprecated; please call lmer() directly

#> Warning: please use glmerControl() instead of lmerControl()

#> Warning: calling glmer() with family=gaussian (identity link) as a shortcut to lmer() is deprecated; please call lmer() directly

#> Warning: please use glmerControl() instead of lmerControl()

#> Warning: calling glmer() with family=gaussian (identity link) as a shortcut to lmer() is deprecated; please call lmer() directly

#> # A tibble: 24 x 8
#>     year effect term             estimate std.error statistic conf.low conf.high
#>    <int> <chr>  <chr>               <dbl>     <dbl>     <dbl>    <dbl>     <dbl>
#>  1  1952 fixed  (Intercept)         0.221     0.280     0.789  -0.328      0.770
#>  2  1952 fixed  scale(gdpPercap)    0.480     0.136     3.54    0.215      0.746
#>  3  1992 fixed  (Intercept)         0.124     0.199     0.623  -0.266      0.514
#>  4  1992 fixed  scale(gdpPercap)    0.653     0.180     3.63    0.300      1.01 
#>  5  2007 fixed  (Intercept)         0.289     0.187     1.54   -0.0784     0.656
#>  6  2007 fixed  scale(gdpPercap)    0.574     0.204     2.82    0.175      0.973
#>  7  1987 fixed  (Intercept)         0.318     0.291     1.09   -0.252      0.888
#>  8  1987 fixed  scale(gdpPercap)    0.537     0.177     3.04    0.190      0.883
#>  9  1977 fixed  (Intercept)         0.573     0.142     4.03    0.294      0.852
#> 10  1977 fixed  scale(gdpPercap)    1.22      0.553     2.20    0.134      2.30 
#> # ... with 14 more rows