Function to run linear model (lm) across multiple grouping variables.

Function to run linear model (lm) across multiple grouping variables.

grouped_lm(data, grouping.vars, formula, output = "tidy")

Arguments

data
Dataframe from which variables are to be taken.

grouping.vars
List of grouping variables.

formula
an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted. The details of model specification are given under ‘Details’.

output
A character describing what output is expected. Two possible options: "tidy" (default), which will return the results, or "glance", which will return model summaries.

Value

A tibble dataframe with tidy results from linear model.

See also

Examples

# loading needed libraries
library(ggplot2)

# getting tidy output of results
grouped_lm(
  data = mtcars,
  grouping.vars = cyl,
  formula = mpg ~ am*wt,
  output = "tidy"
)

#> # A tibble: 12 x 9
#>      cyl term  estimate std.error t.value conf.low conf.high p.value
#>    <dbl> <chr>    <dbl>     <dbl>   <dbl>    <dbl>     <dbl>   <dbl>
#>  1     6 (Int~   63.6      14.1    4.51      18.7    109.    2.04e-2
#>  2     6 am     -41.4      19.0   -2.18    -102.      19.1   1.17e-1
#>  3     6 wt     -13.1       4.16  -3.16     -26.4      0.113 5.11e-2
#>  4     6 am:wt   12.5       6.22   2.02      -7.26    32.4   1.37e-1
#>  5     4 (Int~   13.9      16.1    0.865    -24.1     51.9   4.16e-1
#>  6     4 am      30.3      17.2    1.77     -10.3     70.9   1.21e-1
#>  7     4 wt       3.07      5.44   0.564     -9.79    15.9   5.90e-1
#>  8     4 am:wt  -11.0       6.16  -1.78     -25.5      3.61  1.18e-1
#>  9     8 (Int~   25.1       3.51   7.14      17.2     32.9   3.15e-5
#> 10     8 am      -2.92     25.9   -0.113    -60.5     54.7   9.12e-1
#> 11     8 wt      -2.44      0.842 -2.90      -4.32    -0.563 1.59e-2
#> 12     8 am:wt    0.439     7.63   0.0575   -16.6     17.4   9.55e-1
#> # ... with 1 more variable: significance <chr>

# getting model summaries
# diamonds dataset from ggplot2
grouped_lm(
  data = diamonds,
  grouping.vars = c(cut, color),
  formula = price ~ carat * clarity,
  output = "glance"
)

#> # A tibble: 35 x 14
#>    cut   color r.squared adj.r.squared sigma statistic    df  logLik    AIC
#>    <ord> <ord>     <dbl>         <dbl> <dbl>     <dbl> <int>   <dbl>  <dbl>
#>  1 Ideal E         0.931         0.931  776.     3516.    16 -31501. 63036.
#>  2 Prem~ E         0.927         0.927 1028.     1968.    16 -19516. 39066.
#>  3 Good  E         0.927         0.926  905.      781.    16  -7667. 15369.
#>  4 Prem~ I         0.934         0.934 1303.     1338.    16 -12260. 24554.
#>  5 Good  J         0.946         0.943  885.      363.    15  -2511.  5054.
#>  6 Very~ J         0.957         0.957  862.      994.    16  -5537. 11108.
#>  7 Very~ I         0.946         0.945 1100.     1378.    16 -10132. 20297.
#>  8 Very~ H         0.940         0.939 1032.     1880.    16 -15237. 30508.
#>  9 Fair  E         0.917         0.912  883.      179.    14  -1830.  3690.
#> 10 Ideal J         0.955         0.955  953.     1259.    16  -7409. 14852.
#> # ... with 25 more rows, and 5 more variables: BIC <dbl>, deviance <dbl>,
#> #   df.residual <int>, p.value <dbl>, significance <chr>