Function to run linear mixed-effects model (lmer) across multiple grouping variables.

Function to run linear mixed-effects model (lmer) across multiple grouping variables.

grouped_lmer(data, grouping.vars, formula, REML = TRUE,
  control = lme4::lmerControl(optimizer = "bobyqa", restart_edge = TRUE,
  boundary.tol = 1e-07, calc.derivs = FALSE, use.last.params = FALSE,
  optCtrl = list(maxfun = 2e+09)), p.kr = FALSE, output = "tidy")

Arguments

data
Dataframe from which variables are to be taken.

grouping.vars
List of grouping variables.

formula
a two-sided linear formula object describing both the fixed-effects and random-effects part of the model, with the response on the left of a ~ operator and the terms, separated by + operators, on the right. Random-effects terms are distinguished by vertical bars (|) separating expressions for design matrices from grouping factors. Two vertical bars (||) can be used to specify multiple uncorrelated random effects for the same grouping variable. (Because of the way it is implemented, the ||-syntax works only for design matrices containing numeric (continuous) predictors; to fit models with independent categorical effects, see dummy or the lmer_alt function from the afex package.)

REML
logical scalar - Should the estimates be chosen to optimize the REML criterion (as opposed to the log-likelihood)?

control
a list (of correct class, resulting from lmerControl() or glmerControl() respectively) containing control parameters, including the nonlinear optimizer to be used and parameters to be passed through to the nonlinear optimizer, see the *lmerControl documentation for details.

p.kr
Logical, if TRUE, the computation of p-values is based on conditional F-tests with Kenward-Roger approximation for the df (see 'Details').

output
A character describing what output is expected. Two possible options: "tidy" (default), which will return the results, or "glance", which will return model summaries.

Value

A tibble dataframe with tidy results from linear model or model summaries.

Examples

# loading libraries containing data
library(ggplot2)
library(gapminder)

# getting tidy output of results
# let's use only 50% data to speed it up
groupedstats::grouped_lmer(
  data = dplyr::sample_frac(gapminder, size = 0.5),
  formula = scale(lifeExp) ~ scale(gdpPercap) + (gdpPercap | continent),
  grouping.vars = year,
  output = "tidy"
)

#> Computing p-values via Wald-statistics approximation (treating t as Wald z).

#> Computing p-values via Wald-statistics approximation (treating t as Wald z).

#> Computing p-values via Wald-statistics approximation (treating t as Wald z).

#> Computing p-values via Wald-statistics approximation (treating t as Wald z).

#> Computing p-values via Wald-statistics approximation (treating t as Wald z).

#> Computing p-values via Wald-statistics approximation (treating t as Wald z).

#> Computing p-values via Wald-statistics approximation (treating t as Wald z).

#> Computing p-values via Wald-statistics approximation (treating t as Wald z).

#> Computing p-values via Wald-statistics approximation (treating t as Wald z).

#> Computing p-values via Wald-statistics approximation (treating t as Wald z).

#> Computing p-values via Wald-statistics approximation (treating t as Wald z).

#> Computing p-values via Wald-statistics approximation (treating t as Wald z).

#> # A tibble: 24 x 10
#>     year effect term  estimate std.error t.value conf.low conf.high p.value
#>    <int> <chr>  <chr>    <dbl>     <dbl>   <dbl>    <dbl>     <dbl>   <dbl>
#>  1  1952 fixed  (Int~    0.220     0.382   0.575  -0.529      0.969 5.65e-1
#>  2  1952 fixed  scal~    0.904     0.532   1.70   -0.138      1.95  8.91e-2
#>  3  2007 fixed  (Int~    0.296     0.204   1.45   -0.105      0.696 1.48e-1
#>  4  2007 fixed  scal~    0.496     0.214   2.32    0.0773     0.916 2.03e-2
#>  5  1992 fixed  (Int~    0.414     0.116   3.58    0.187      0.641 3.47e-4
#>  6  1992 fixed  scal~    0.782     0.375   2.09    0.0471     1.52  3.70e-2
#>  7  1987 fixed  (Int~    0.313     0.248   1.26   -0.173      0.798 2.07e-1
#>  8  1987 fixed  scal~    0.590     0.190   3.11    0.218      0.963 1.90e-3
#>  9  1962 fixed  (Int~    0.321     0.326   0.985  -0.318      0.961 3.25e-1
#> 10  1962 fixed  scal~    0.897     0.400   2.24    0.114      1.68  2.48e-2
#> # ... with 14 more rows, and 1 more variable: significance <chr>

# getting model summaries
# let's use only 50% data to speed it up
grouped_lmer(
  data = ggplot2::diamonds,
  formula = scale(price) ~ scale(carat) + (carat | color),
  REML = FALSE,
  grouping.vars = c(cut, clarity),
  output = "glance"
)

#> # A tibble: 40 x 8
#>    cut       clarity sigma logLik    AIC    BIC deviance df.residual
#>    <ord>     <ord>   <dbl>  <dbl>  <dbl>  <dbl>    <dbl>       <int>
#>  1 Ideal     SI2     0.252 -134.   279.   315.     267.         2592
#>  2 Premium   SI1     0.254 -201.   414.   451.     402.         3569
#>  3 Good      VS1     0.269  -83.1  178.   205.     166.          642
#>  4 Premium   VS2     0.243  -47.4  107.   144.      94.8        3351
#>  5 Good      SI2     0.273 -151.   313.   343.     301.         1075
#>  6 Very Good VVS2    0.232   27.9  -43.9  -13.1    -55.9        1229
#>  7 Very Good VVS1    0.245  -31.1   74.2  102.      62.2         783
#>  8 Very Good SI1     0.256 -204.   420.   456.     408.         3234
#>  9 Fair      VS2     0.287  -59.8  132.   153.     120.          255
#> 10 Very Good VS1     0.246  -52.4  117.   150.     105.         1769
#> # ... with 30 more rows