Function to run simple linear regression (slr) on multiple variables across multiple grouping variables.

Function to run simple linear regression (slr) on multiple variables across multiple grouping variables.

grouped_slr(data, dep.vars, indep.vars, grouping.vars)

Arguments

data
Dataframe from which variables are to be taken.

dep.vars
List criterion or dependent variables for simple linear model (y in y ~ x).

indep.vars
List predictor or independent variables for simple linear model (x in y ~ x).

grouping.vars
List of grouping variables.

Value

A tibble dataframe with tidy results from simple linear regression analyses. The esimates are standardized, i.e. the lm model used is scale(y) ~ scale(x), and not y ~ x.

Examples

# in case of just one grouping variable
groupedstats::grouped_slr(data = iris,
dep.vars = c(Sepal.Length, Petal.Length),
indep.vars = c(Sepal.Width, Petal.Width),
grouping.vars = Species)

#> # A tibble: 6 x 9
#>   Species  formula       t.value estimate conf.low conf.high std.error  p.value
#>   <fct>    <chr>           <dbl>    <dbl>    <dbl>     <dbl>     <dbl>    <dbl>
#> 1 setosa   Sepal.Length~    7.68    0.743   0.548      0.937    0.0967 6.71e-10
#> 2 versico~ Sepal.Length~    4.28    0.526   0.279      0.773    0.123  8.77e- 5
#> 3 virgini~ Sepal.Length~    3.56    0.457   0.199      0.715    0.128  8.43e- 4
#> 4 setosa   Petal.Length~    2.44    0.332   0.0578     0.605    0.136  1.86e- 2
#> 5 versico~ Petal.Length~    8.83    0.787   0.607      0.966    0.0891 1.27e-11
#> 6 virgini~ Petal.Length~    2.36    0.322   0.0474     0.597    0.137  2.25e- 2
#> # ... with 1 more variable: significance <chr>

# in case of multiple grouping variables
groupedstats::grouped_slr( data = mtcars,
dep.vars = c(wt, mpg),
indep.vars = c(drat, disp),
grouping.vars = c(am, cyl))

#> Warning: NaNs produced

#> Warning: essentially perfect fit: summary may be unreliable

#> Warning: essentially perfect fit: summary may be unreliable

#> Warning: NaNs produced

#> # A tibble: 12 x 10
#>       am   cyl formula  t.value estimate conf.low conf.high std.error   p.value
#>    <dbl> <dbl> <chr>      <dbl>    <dbl>    <dbl>     <dbl>     <dbl>     <dbl>
#>  1     1     6 wt ~ d~   -0.102   -0.101  -12.7     12.5        0.995   0.935  
#>  2     1     4 wt ~ d~   -0.568   -0.226   -1.20     0.747      0.398   0.591  
#>  3     0     6 wt ~ d~    0.456    0.307   -2.59     3.20       0.673   0.693  
#>  4     0     8 wt ~ d~   -0.379   -0.119   -0.819    0.581      0.314   0.713  
#>  5     0     4 wt ~ d~    0.466    0.422  -11.1     11.9        0.906   0.722  
#>  6     1     8 wt ~ d~  NaN       -1      NaN      NaN        NaN     NaN      
#>  7     1     6 mpg ~ ~  Inf        1.       1.       1.         0       0      
#>  8     1     4 mpg ~ ~   -3.72    -0.835   -1.38    -0.285      0.225   0.00991
#>  9     0     6 mpg ~ ~    1.28     0.670   -1.59     2.93       0.525   0.330  
#> 10     0     8 mpg ~ ~   -2.00    -0.535   -1.13     0.0598     0.267   0.0729 
#> 11     0     4 mpg ~ ~    2.57     0.932   -3.67     5.53       0.362   0.236  
#> 12     1     8 mpg ~ ~  NaN        1      NaN      NaN        NaN     NaN      
#> # ... with 1 more variable: significance <chr>