Function to get descriptive statistics for multiple variables for all grouping variable levels

Function to get descriptive statistics for multiple variables for all grouping variable levels

grouped_summary(data, grouping.vars, measures = NULL,
  measures.type = "numeric", topcount.long = FALSE)

Arguments

data
Dataframe from which variables need to be taken.

grouping.vars
A list of grouping variables.

measures
List variables for which summary needs to computed. If not specified, all variables of type specified in the argument measures.type will be used to calculate summaries. Don't explicitly set measures.type = NULL in function call, which will produce an error because the function will try to find a column in a dataframe named "NULL".

measures.type
A character indicating whether summary for numeric ("numeric") or factor/character ("factor") variables is expected (Default: measures.type = "numeric"). This function can't be used for both numeric and variables simultaneously.

topcount.long
If measures.type = factor, you can get the top counts in long format for plotting purposes. (Default: topcount.long = FALSE).

Value

Dataframe with descriptive statistics for numeric variables (n, mean, sd, median, min, max)

Examples

# another possibility
groupedstats::grouped_summary(
  data = datasets::iris,
  grouping.vars = Species,
  measures = Sepal.Length:Petal.Width,
  measures.type = "numeric"
)

#> [90m# A tibble: 12 x 16[39m
#>    Species type  variable missing complete     n  mean    sd   min   p25 median
#>    [3m[90m<fct>[39m[23m   [3m[90m<chr>[39m[23m [3m[90m<chr>[39m[23m      [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m  [3m[90m<dbl>[39m[23m
#> [90m 1[39m setosa  nume~ Petal.L~       0       50    50  1.46  0.17   1    1.4    1.5 
#> [90m 2[39m setosa  nume~ Petal.W~       0       50    50  0.25  0.11   0.1  0.2    0.2 
#> [90m 3[39m setosa  nume~ Sepal.L~       0       50    50  5.01  0.35   4.3  4.8    5   
#> [90m 4[39m setosa  nume~ Sepal.W~       0       50    50  3.43  0.38   2.3  3.2    3.4 
#> [90m 5[39m versic~ nume~ Petal.L~       0       50    50  4.26  0.47   3    4      4.35
#> [90m 6[39m versic~ nume~ Petal.W~       0       50    50  1.33  0.2    1    1.2    1.3 
#> [90m 7[39m versic~ nume~ Sepal.L~       0       50    50  5.94  0.52   4.9  5.6    5.9 
#> [90m 8[39m versic~ nume~ Sepal.W~       0       50    50  2.77  0.31   2    2.52   2.8 
#> [90m 9[39m virgin~ nume~ Petal.L~       0       50    50  5.55  0.55   4.5  5.1    5.55
#> [90m10[39m virgin~ nume~ Petal.W~       0       50    50  2.03  0.27   1.4  1.8    2   
#> [90m11[39m virgin~ nume~ Sepal.L~       0       50    50  6.59  0.64   4.9  6.23   6.5 
#> [90m12[39m virgin~ nume~ Sepal.W~       0       50    50  2.97  0.32   2.2  2.8    3   
#> [90m# ... with 5 more variables: p75 [3m[90m<dbl>[90m[23m, max [3m[90m<dbl>[90m[23m, std.error [3m[90m<dbl>[90m[23m,
#> #   mean.low.conf [3m[90m<dbl>[90m[23m, mean.high.conf [3m[90m<dbl>[90m[23m[39m

# if you have just one variable per argument, you need not use `c()`
groupedstats::grouped_summary(
  data = datasets::ToothGrowth,
  grouping.vars = supp,
  measures = len,
  measures.type = "numeric"
)

#> [90m# A tibble: 2 x 16[39m
#>   supp  type  variable missing complete     n  mean    sd   min   p25 median
#>   [3m[90m<fct>[39m[23m [3m[90m<chr>[39m[23m [3m[90m<chr>[39m[23m      [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m  [3m[90m<dbl>[39m[23m
#> [90m1[39m VC    nume~ len            0       30    30  17.0  8.27   4.2  11.2   16.5
#> [90m2[39m OJ    nume~ len            0       30    30  20.7  6.61   8.2  15.5   22.7
#> [90m# ... with 5 more variables: p75 [3m[90m<dbl>[90m[23m, max [3m[90m<dbl>[90m[23m, std.error [3m[90m<dbl>[90m[23m,
#> #   mean.low.conf [3m[90m<dbl>[90m[23m, mean.high.conf [3m[90m<dbl>[90m[23m[39m