Descriptive statistics for multiple variables for all grouping variable levels

Descriptive statistics for multiple variables for all grouping variable levels

grouped_summary(
  data,
  grouping.vars,
  measures = NULL,
  measures.type = "numeric",
  topcount.long = FALSE,
  k = 2L,
  ...
)

Arguments

data
Dataframe from which variables need to be taken.

grouping.vars
A list of grouping variables.

measures
List variables for which summary needs to computed. If not specified, all variables of type specified in the argument measures.type will be used to calculate summaries. Don't explicitly set measures.type = NULL in function call, which will produce an error because the function will try to find a column in a dataframe named "NULL".

measures.type
A character indicating whether summary for numeric ("numeric") or factor/character ("factor") variables is expected (Default: measures.type = "numeric"). This function can't be used for both numeric and variables simultaneously.

topcount.long
If measures.type = factor, you can get the top counts in long format for plotting purposes. (Default: topcount.long = FALSE).

k
Number of digits after decimal point (should be an integer) (Default: k = 3).

...
Currently ignored.

Value

Dataframe with descriptive statistics for numeric variables (n, mean, sd, median, min, max)

Examples

# for reproducibility
set.seed(123)

# another possibility
groupedstats::grouped_summary(
  data = iris,
  grouping.vars = Species,
  measures = Sepal.Length:Petal.Width,
  measures.type = "numeric"
)

#> Joining, by = "Species"

#> # A tibble: 12 x 16
#>    Species type  variable missing complete  mean    sd   min   p25 median   p75
#>    <fct>   <chr> <chr>      <int>    <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl>
#>  1 setosa  nume~ Sepal.L~       0        1 5.01  0.352   4.3  4.8    5     5.2 
#>  2 setosa  nume~ Sepal.W~       0        1 3.43  0.379   2.3  3.2    3.4   3.68
#>  3 setosa  nume~ Petal.L~       0        1 1.46  0.174   1    1.4    1.5   1.58
#>  4 setosa  nume~ Petal.W~       0        1 0.246 0.105   0.1  0.2    0.2   0.3 
#>  5 versic~ nume~ Sepal.L~       0        1 5.94  0.516   4.9  5.6    5.9   6.3 
#>  6 versic~ nume~ Sepal.W~       0        1 2.77  0.314   2    2.52   2.8   3   
#>  7 versic~ nume~ Petal.L~       0        1 4.26  0.470   3    4      4.35  4.6 
#>  8 versic~ nume~ Petal.W~       0        1 1.33  0.198   1    1.2    1.3   1.5 
#>  9 virgin~ nume~ Sepal.L~       0        1 6.59  0.636   4.9  6.22   6.5   6.9 
#> 10 virgin~ nume~ Sepal.W~       0        1 2.97  0.322   2.2  2.8    3     3.18
#> 11 virgin~ nume~ Petal.L~       0        1 5.55  0.552   4.5  5.1    5.55  5.88
#> 12 virgin~ nume~ Petal.W~       0        1 2.03  0.275   1.4  1.8    2     2.3 
#> # ... with 5 more variables: max <dbl>, n <int>, std.error <dbl>,
#> #   mean.low.conf <dbl>, mean.high.conf <dbl>

# if no measures are chosen, all relevant columns will be summarized
groupedstats::grouped_summary(
  data = ggplot2::msleep,
  grouping.vars = vore,
  measures.type = "factor"
)

#> Joining, by = "vore"

#> # A tibble: 16 x 9
#>    vore   type   variable missing complete ordered n_unique top_counts         n
#>    <fct>  <chr>  <chr>      <int>    <dbl> <lgl>      <int> <chr>          <int>
#>  1 carni  factor name           0    1     FALSE         19 Arc: 1, Bot: ~    19
#>  2 carni  factor genus          0    1     FALSE         16 Pan: 3, Vul: ~    19
#>  3 carni  factor order          0    1     FALSE          6 Car: 12, Cet:~    19
#>  4 carni  factor conserv~       5    0.737 FALSE          6 lc: 5, vu: 4,~    14
#>  5 herbi  factor name           0    1     FALSE         32 Afr: 1, Arc: ~    32
#>  6 herbi  factor genus          0    1     FALSE         29 Spe: 3, Equ: ~    32
#>  7 herbi  factor order          0    1     FALSE          9 Rod: 16, Art:~    32
#>  8 herbi  factor conserv~       6    0.812 FALSE          6 lc: 10, dom: ~    26
#>  9 insec~ factor name           0    1     FALSE          5 Big: 1, Eas: ~     5
#> 10 insec~ factor genus          0    1     FALSE          5 Ept: 1, Myo: ~     5
#> 11 insec~ factor order          0    1     FALSE          4 Chi: 2, Cin: ~     5
#> 12 insec~ factor conserv~       2    0.6   FALSE          2 lc: 2, en: 1,~     3
#> 13 omni   factor name           0    1     FALSE         20 Afr: 1, Afr: ~    20
#> 14 omni   factor genus          0    1     FALSE         20 Aot: 1, Bla: ~    20
#> 15 omni   factor order          0    1     FALSE          8 Pri: 10, Sor:~    20
#> 16 omni   factor conserv~      11    0.450 FALSE          2 lc: 8, dom: 1~     9

# for factors, you can also convert the dataframe to long format with counts
groupedstats::grouped_summary(
  data = ggplot2::msleep,
  grouping.vars = c(vore),
  measures = c(genus:order),
  measures.type = "factor",
  topcount.long = TRUE
)

#> Joining, by = "vore"

#> # A tibble: 32 x 3
#>    vore  factor.level count
#>    <fct> <chr>        <int>
#>  1 carni Pan              3
#>  2 carni Vul              2
#>  3 carni Aci              1
#>  4 carni Cal              1
#>  5 carni Car             12
#>  6 carni Cet              3
#>  7 carni Cin              1
#>  8 carni Did              1
#>  9 herbi Spe              3
#> 10 herbi Equ              2
#> # ... with 22 more rows