Function to get descriptive statistics for multiple variables for all grouping variable levels
Source: R/grouped_summary.R

Function to get descriptive statistics for multiple variables for all grouping variable levels

Arguments

data
Dataframe from which variables need to be taken.

grouping.vars
A list of grouping variables.

measures
List variables for which summary needs to computed.

measures.type
A character indicating whether summary for numeric ("numeric") or factor/character ("factor") variables is expected (Default: measures.type = "numeric"). This function can't be used for both numeric and variables simultaneously.

Value

Dataframe with descriptive statistics for numeric variables (n, mean, sd, median, min, max)

Examples

library(datasets)

# if you have multiple variable for each argument
groupedstats::grouped_summary(
data = datasets::mtcars,
grouping.vars = c(am, cyl),
measures = c(wt, mpg)
)

#> # A tibble: 12 x 14
#>       am   cyl type    variable missing complete     n  mean    sd   min   p25
#>    <dbl> <dbl> <fct>   <fct>      <dbl>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1    1.    6. numeric mpg           0.       3.    3. 20.6  0.750 19.7  20.4 
#>  2    1.    6. numeric wt            0.       3.    3.  2.75 0.130  2.62  2.70
#>  3    1.    4. numeric mpg           0.       8.    8. 28.1  4.48  21.4  25.2 
#>  4    1.    4. numeric wt            0.       8.    8.  2.04 0.410  1.51  1.78
#>  5    0.    6. numeric mpg           0.       4.    4. 19.1  1.63  17.8  18.0 
#>  6    0.    6. numeric wt            0.       4.    4.  3.39 0.120  3.21  3.38
#>  7    0.    8. numeric mpg           0.      12.   12. 15.0  2.77  10.4  14.0 
#>  8    0.    8. numeric wt            0.      12.   12.  4.10 0.770  3.44  3.56
#>  9    0.    4. numeric mpg           0.       3.    3. 22.9  1.45  21.5  22.2 
#> 10    0.    4. numeric wt            0.       3.    3.  2.94 0.410  2.46  2.81
#> 11    1.    8. numeric mpg           0.       2.    2. 15.4  0.570 15.0  15.2 
#> 12    1.    8. numeric wt            0.       2.    2.  3.37 0.280  3.17  3.27
#> # ... with 3 more variables: median <dbl>, p75 <dbl>, max <dbl>

# another possibility
groupedstats::grouped_summary(
data = datasets::iris,
grouping.vars = Species,
measures = Sepal.Length:Petal.Width,
measures.type = "numeric"
)

#> # A tibble: 12 x 13
#>    Species type  variable missing complete     n  mean    sd   min   p25 median
#>    <fct>   <fct> <fct>      <dbl>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>
#>  1 setosa  nume~ Petal.L~      0.      50.   50. 1.46  0.170 1.00  1.40   1.50 
#>  2 setosa  nume~ Petal.W~      0.      50.   50. 0.250 0.110 0.100 0.200  0.200
#>  3 setosa  nume~ Sepal.L~      0.      50.   50. 5.01  0.350 4.30  4.80   5.00 
#>  4 setosa  nume~ Sepal.W~      0.      50.   50. 3.43  0.380 2.30  3.20   3.40 
#>  5 versic~ nume~ Petal.L~      0.      50.   50. 4.26  0.470 3.00  4.00   4.35 
#>  6 versic~ nume~ Petal.W~      0.      50.   50. 1.33  0.200 1.00  1.20   1.30 
#>  7 versic~ nume~ Sepal.L~      0.      50.   50. 5.94  0.520 4.90  5.60   5.90 
#>  8 versic~ nume~ Sepal.W~      0.      50.   50. 2.77  0.310 2.00  2.52   2.80 
#>  9 virgin~ nume~ Petal.L~      0.      50.   50. 5.55  0.550 4.50  5.10   5.55 
#> 10 virgin~ nume~ Petal.W~      0.      50.   50. 2.03  0.270 1.40  1.80   2.00 
#> 11 virgin~ nume~ Sepal.L~      0.      50.   50. 6.59  0.640 4.90  6.23   6.50 
#> 12 virgin~ nume~ Sepal.W~      0.      50.   50. 2.97  0.320 2.20  2.80   3.00 
#> # ... with 2 more variables: p75 <dbl>, max <dbl>

# if you have just one variable per argument, you can also not use `c()`
groupedstats::grouped_summary(
data = datasets::ToothGrowth,
grouping.vars = supp,
measures = len,
measures.type = "numeric"
)

#> # A tibble: 2 x 13
#>   supp  type    variable missing complete     n  mean    sd   min   p25 median
#>   <fct> <fct>   <fct>      <dbl>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>
#> 1 VC    numeric len           0.      30.   30.  17.0  8.27  4.20  11.2   16.5
#> 2 OJ    numeric len           0.      30.   30.  20.7  6.61  8.20  15.5   22.7
#> # ... with 2 more variables: p75 <dbl>, max <dbl>