Function to get descriptive statistics for multiple variables for all grouping variable levels

Function to get descriptive statistics for multiple variables for all grouping variable levels

grouped_summary(data, grouping.vars, measures = NULL,
  measures.type = "numeric", topcount.long = FALSE)

Arguments

data
Dataframe from which variables need to be taken.

grouping.vars
A list of grouping variables.

measures
List variables for which summary needs to computed. If not specified, all variables of type specified in the argument measures.type will be used to calculate summaries. Don't explicitly set measures.type = NULL in function call, which will produce an error because the function will try to find a column in a dataframe named "NULL".

measures.type
A character indicating whether summary for numeric ("numeric") or factor/character ("factor") variables is expected (Default: measures.type = "numeric"). This function can't be used for both numeric and variables simultaneously.

topcount.long
If measures.type = factor, you can get the top counts in long format for plotting purposes. (Default: topcount.long = FALSE).

Value

Dataframe with descriptive statistics for numeric variables (n, mean, sd, median, min, max)

Examples

# another possibility
groupedstats::grouped_summary(
  data = datasets::iris,
  grouping.vars = Species,
  measures = Sepal.Length:Petal.Width,
  measures.type = "numeric"
)

#> # A tibble: 12 x 16
#>    Species variable type  missing complete     n  mean    sd   min   p25 median
#>    <fct>   <chr>    <chr>   <dbl>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>
#>  1 setosa  Sepal.L~ nume~       0       50    50 5.01  0.352   4.3  4.8    5   
#>  2 setosa  Sepal.W~ nume~       0       50    50 3.43  0.379   2.3  3.2    3.4 
#>  3 setosa  Petal.L~ nume~       0       50    50 1.46  0.174   1    1.4    1.5 
#>  4 setosa  Petal.W~ nume~       0       50    50 0.246 0.105   0.1  0.2    0.2 
#>  5 versic~ Sepal.L~ nume~       0       50    50 5.94  0.516   4.9  5.6    5.9 
#>  6 versic~ Sepal.W~ nume~       0       50    50 2.77  0.314   2    2.52   2.8 
#>  7 versic~ Petal.L~ nume~       0       50    50 4.26  0.470   3    4      4.35
#>  8 versic~ Petal.W~ nume~       0       50    50 1.33  0.198   1    1.2    1.3 
#>  9 virgin~ Sepal.L~ nume~       0       50    50 6.59  0.636   4.9  6.22   6.5 
#> 10 virgin~ Sepal.W~ nume~       0       50    50 2.97  0.322   2.2  2.8    3   
#> 11 virgin~ Petal.L~ nume~       0       50    50 5.55  0.552   4.5  5.1    5.55
#> 12 virgin~ Petal.W~ nume~       0       50    50 2.03  0.275   1.4  1.8    2   
#> # ... with 5 more variables: p75 <dbl>, max <dbl>, std.error <dbl>,
#> #   mean.low.conf <dbl>, mean.high.conf <dbl>

# if you have just one variable per argument, you need not use `c()`
groupedstats::grouped_summary(
  data = datasets::ToothGrowth,
  grouping.vars = supp,
  measures = len,
  measures.type = "numeric"
)

#> # A tibble: 2 x 16
#>   supp  variable type  missing complete     n  mean    sd   min   p25 median
#>   <fct> <chr>    <chr>   <dbl>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>
#> 1 VC    len      nume~       0       30    30  17.0  8.27   4.2  11.2   16.5
#> 2 OJ    len      nume~       0       30    30  20.7  6.61   8.2  15.5   22.7
#> # ... with 5 more variables: p75 <dbl>, max <dbl>, std.error <dbl>,
#> #   mean.low.conf <dbl>, mean.high.conf <dbl>