Function to run two-sample Wilcoxon tests on multiple variables across multiple grouping variables.

Function to run two-sample Wilcoxon tests on multiple variables across multiple grouping variables.

grouped_wilcox(data, dep.vars, indep.vars, grouping.vars, paired = FALSE,
  correct = TRUE)

Arguments

data
Dataframe from which variables are to be taken.

dep.vars
List dependent variables for a two-sample Wilcoxon tests (y in y ~ x).

indep.vars
List independent variables for a two-sample Wilcoxon tests (x in y ~ x).

grouping.vars
List of grouping variables.

paired
A logical indicating whether you want a paired two-sample Wilcoxon tests (Default: paired = FALSE).

correct
A logical indicating whether to apply continuity correction in the normal approximation for the p-value (Default: correct = TRUE).

Value

A tibble dataframe with tidy results from two-sample Wilcoxon tests analyses.

Examples

# only with one grouping variable
groupedstats::grouped_wilcox(
  data = dplyr::filter(.data = ggplot2::diamonds, color == "E" | color == "J"),
  dep.vars = depth:price,
  indep.vars = color,
  grouping.vars = clarity,
  paired = FALSE
)

#> # A tibble: 16 x 9
#>    clarity formula  method      statistic estimate  conf.low conf.high  p.value
#>    <ord>   <chr>    <fct>           <dbl>    <dbl>     <dbl>     <dbl>    <dbl>
#>  1 SI2     depth ~~ Wilcoxon r~   380600  -2.00e-1  -3.00e-1  -3.64e-5 1.54e- 2
#>  2 SI1     depth ~~ Wilcoxon r~   918892   3.93e-5 -10.00e-2   1.00e-1 6.77e- 1
#>  3 VS1     depth ~~ Wilcoxon r~   276743  -4.00e-1  -5.00e-1  -3.00e-1 7.03e-12
#>  4 VVS2    depth ~~ Wilcoxon r~    52618  -4.00e-1  -6.00e-1  -2.00e-1 4.18e- 4
#>  5 VS2     depth ~~ Wilcoxon r~   816784. -2.00e-1  -3.00e-1 -10.00e-2 8.86e- 5
#>  6 I1      depth ~~ Wilcoxon r~     1570. -2.00e+0  -3.00e+0 -10.00e-1 1.21e- 4
#>  7 VVS1    depth ~~ Wilcoxon r~    19455  -3.00e-1  -5.00e-1 -10.00e-2 5.07e- 3
#>  8 IF      depth ~~ Wilcoxon r~     3286  -3.00e-1  -7.00e-1  -2.51e-5 4.79e- 2
#>  9 SI2     price ~~ Wilcoxon r~   269016  -1.84e+3  -2.18e+3  -1.50e+3 8.82e-31
#> 10 SI1     price ~~ Wilcoxon r~   608724. -1.51e+3  -1.79e+3  -1.31e+3 8.20e-43
#> 11 VS1     price ~~ Wilcoxon r~   277389  -1.10e+3  -1.41e+3  -7.45e+2 1.11e-11
#> 12 VVS2    price ~~ Wilcoxon r~    42795  -1.76e+3  -2.92e+3  -1.20e+3 2.23e-10
#> 13 VS2     price ~~ Wilcoxon r~   560930  -1.89e+3  -2.23e+3  -1.54e+3 1.08e-54
#> 14 I1      price ~~ Wilcoxon r~     2017  -1.15e+3  -1.83e+3  -7.90e+1 3.68e- 2
#> 15 VVS1    price ~~ Wilcoxon r~    19086  -3.62e+2  -8.59e+2  -1.07e+2 2.57e- 3
#> 16 IF      price ~~ Wilcoxon r~     5085   2.96e+2   1.25e+2   4.49e+2 4.94e- 3
#> # ... with 1 more variable: significance <chr>